Microsoft and City of Austin Scripts Combined
Enabled the Microsoft and City of Austin scripts to both be executed when the Job_Board_Scripts_Combined.py file is ran. Had to adjust the names of the files and the way the code is formatted to do this.
