+++ .gitignore
@@ -2,3 +2,5 @@
 /MSFT_JOBS
 .env
 MSFT_Web_Scraping.py
+__pycache__/CityofAustin.cpython-312.pyc
+__pycache__/MSFT.cpython-312.pyc
